pyproject.toml
[tool.poetry]
name = "alphabet counter bot thing"
version = "0.1.0"
description = "A alphabet counter bot"
authors = ["timelessnesses <mooping3roblox@gmail.com>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
orjson = "^3.6.8"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
asyncpg = "^0.29.0"
python-dateutil = "^2.8.2"
parsedatetime = "^2.6"
python-dotenv = "^0.20.0"
jishaku = "^2.5.0"
psutil = "^5.9.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
watchdog = "^2.1.7"
Flask = "^2.1.2"
uvloop = {version = "^0.16.0", platform = "linux"}
asyncpg-stubs = "^0.29.1"

[tool.poetry.dev-dependencies]
watchdog = "^2.1.7"
black = "^22.3.0"
isort = "^5.10.1"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
